requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
langchain>=0.2.14
langchain-groq>=0.1.5
streamlit>=1.37.0
pandas>=2.2.2